refactor: change location of pages
