+++ src/style/_footer.scss
@@ -23,7 +23,7 @@
     gap: 10px;
   }
 
-  &__mixer-image {
+  &__mixer {
     width: 4rem;
     height: 4rem;
     background-color: transparent;
